requirements.txt
# requirements.txt
fastapi
uvicorn[standard]
scikit-learn
numpy
dagshub
mlflow-skinny
requests